try bazel
I haven't gone as far as I wanted to, but i've 
ran into a couple of issues:

standing issues:

1. Bazel distributions in nixpkgs for some reason
cause rules_xcodeproj to break when generating
an xcode project. It's unable to grab a path to
the working directory of the "output_base".

2. We'll have to write our own rules to get 
toolchains working with nix at least. It'd be a
tad bit confusing on who we let manage the
dependencies essentially. rules_nixpkgs aren't
as good as I hoped them to be unfortunately.

3. rules_xcodeproj does not support SwiftUI 
Previews in XCode 16, I don't know what else
it doesn't support. 

good things:

1. it's pretty fast, with great LSP support (starpls)
and has a vast community backing it unlike buck2.

2. JetBrains has experimental first-class support 
for Bazel in Android Studio. Which is great,
similar to their other editors, so editor support
works out of the box in few languages atm with 
Bazel. See https://github.com/JetBrains/hirschgarten/blob/main/docs/feature_matrix.md  

verdict:

I don't know yet, it's only bound to get better
so our problem is mostly nix or bazel. Until 
of course we get a solution that doesn't
force us to use an absolute path like /nix/store.

It's a lot more hopeful than Buck2. Which still
doesn't have a proper LSP, open-source Apple
and Android rules, and local sandboxing.